requirements.txt
Flask
Flask-Cors
opencv-python
numpy
onnxruntime
pyserial
